Moving fp16/bf16 conversion methods to OrtUtil.
